FLUX: move dataset to single file
